modified cls structure init room helper
